todos with context upd
