+++ src/styles/todo-list.scss
@@ -71,6 +71,7 @@
   }
 
   &__title-field {
+    box-sizing: border-box;
     width: 100%;
     padding: 11px 14px;
