+++ src/styles/todoapp.scss
@@ -56,6 +56,7 @@
   }
 
   &__new-todo {
+    box-sizing: border-box;
     width: 100%;
     padding: 16px 16px 16px 60px;
