+++ src/types/filters.ts
@@ -1,0 +1,1 @@
+export type FilterType = 'all' | 'active' | 'completed';
