+++ src/types/state.ts
@@ -1,0 +1,7 @@
+import { FilterType } from './filters';
+import { Todo } from './todo';
+
+export type State = {
+  todos: Todo[];
+  filter: FilterType;
+};
